+++ packages/micro-generators/code-quality/src/index.ts
@@ -1,5 +1,5 @@
-import { AppType } from '@test-release/core'
-import { createEditorConfig } from '@test-release/editor-config'
+import { AppType, builder } from '@test-release/core'
+import { createEditorConfigSchema } from '@test-release/editor-config'
 import { createPrettierConfig } from '@test-release/prettier'
 import { createStylelintWebConfig, createStylelintMobileConfig } from '@test-release/stylelint'
 
@@ -8,40 +8,18 @@ export interface IOptions {
 }
 
 export const createCodeQualityConfig = ({ appType }: IOptions) => {
+  const schema = builder('codequality')
   const hasPrettier = true
 
-  const editorConfigSchema = createEditorConfig()
+  const editorConfigSchema = createEditorConfigSchema()
   const prettierSchema = createPrettierConfig({ appType })
   const stylelintConfig = appType === AppType.MOBILE
     ? createStylelintWebConfig({ hasPrettier })
     : createStylelintMobileConfig({ hasPrettier })
 
-  return {
-    name: 'code-quality',
-    sources: [
-      {
-        name: 'editorconfig',
-        source: editorConfigSchema.source,
-      },
-      {
-        name: 'prettier',
-        source: prettierSchema.source,
-      },
-      {
-        name: 'stylelint',
-        source: stylelintConfig.source,
-        context: stylelintConfig.context,
-      },
-    ],
-    packageJson: {
-      scripts: {
-        ...prettierSchema.packageJson.scripts,
-        ...stylelintConfig.packageJson.scripts,
-      },
-      devDependencies: [
-        ...prettierSchema.packageJson.devDependencies,
-        ...stylelintConfig.packageJson.devDependencies,
-      ]
-    }
-  }
+  schema.combineSchema(editorConfigSchema)
+  schema.combineSchema(prettierSchema)
+  schema.combineSchema(stylelintConfig)
+
+  return schema.toJson()
 }