+++ packages/micro-generators/code-quality/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@test-release/code-quality-v1.1.0](https://github.com/developer239/test-release/compare/@test-release/code-quality-v1.0.0...@test-release/code-quality-v1.1.0) (2020-05-18)
+
+
+### Features
+
+* update code quality package ([dd6323f](https://github.com/developer239/test-release/commit/dd6323f1898e0eeea9322fbc28ae697cb3e6770b))
+
 # @test-release/code-quality-v1.0.0 (2020-05-13)